% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKS.R
\name{plotKS}
\alias{plotKS}
\title{KS Plot & Statistic}
\usage{
plotKS(a, b, la = "Distribution A", lb = "Distribution B")
}
\arguments{
\item{a}{(numeric) vector of scores for the first class}

\item{b}{(numeric) vector of scores for the second class}

\item{la}{(character) legend label for the first class}

\item{lb}{(character) legend label for the second class}
}
\value{
a list containing the following elements:
\item{ks.stat} {KS statistic}
\item{ks.plot} {a ggplot2 object for the KS plot}
}
\description{
Plot the cumulative distributions of each class by a score variable and calculate the KS
statistic denoting the max vertical distance between the two cumulative distributions.
}
\examples{
a <- rnorm(5000, 0, 2)
b <- rnorm(5000, 1, 3)
res <- plotKS(a, b, la = 'Non-Defaults', lb = 'Defaults')
res$ks.stat
res$ks.plot

# compare with internal stats function
ks.test(a, b)
}

