% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsSummary.R
\name{tsSummary}
\alias{tsSummary}
\title{Calculate summary statistics for time series variables
For each combination of [id/measure] summary statistics are calculated over all periods}
\usage{
tsSummary(df, idvar, pfx, sfx, ...)
}
\arguments{
\item{df}{(data frame) data set containing the time series variables}

\item{idvar}{(character) name of the variable that uniquely identifies subjects in the data}

\item{pfx}{(character) vector of prefix patterns for each time series variable}

\item{sfx}{(character) pattern yielding period-specific values for each time series variable}

\item{...}{(functions) functions to evaluate for each [id/measure] (built-in or user-defined)}
}
\value{
a data frame containing the \code{idvar} and one column for each combination of [measure/stat]
}
\description{
Calculate summary statistics for time series variables
For each combination of [id/measure] summary statistics are calculated over all periods
}

