% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsTrends.R
\name{tsTrends}
\alias{tsTrends}
\title{Calculate linear trends for a set of time series variables
For each combination of \code{idvar} and \code{measure} regress measure values against time period.
The result will be the best linear approximation of the measure slope as a function of time}
\usage{
tsTrends(df, idvar, pfx, sfx, reverse = FALSE)
}
\arguments{
\item{df}{(data frame) data frame containing the time series variables of interest}

\item{idvar}{(character) name of the variable that uniquely identifies observations in the data}

\item{pfx}{(character) vector of prefix patterns for each time series measure}

\item{sfx}{(character) pattern to match time period suffix values}

\item{reverse}{(logical) reverse the order of period values when calculating linear trends}
}
\value{
a data frame containing the \code{idvar} and one column for each time series measure containing the 
linear trends for that \code{idvar} and \code{measure} combination
}
\description{
Calculate linear trends for a set of time series variables
For each combination of \code{idvar} and \code{measure} regress measure values against time period.
The result will be the best linear approximation of the measure slope as a function of time
}

