% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varCluster.R
\name{varCluster}
\alias{varCluster}
\title{Predictor Clustering and Cluster Summary Variable Extraction}
\usage{
varCluster(x, y, corr.min = sqrt(2)/2, clus.summary = "max.pw",
  corr.method = "pearson", corr.use = "complete.obs",
  clus.method = "complete")
}
\arguments{
\item{x}{(data frame) data frame containing the set of numeric predictors to be clustered}

\item{y}{(numeric) the intended response variable (to generate x-y correlations)}

\item{corr.min}{(numeric) minimum correlation required to join variables into clusters}

\item{clus.summary}{(character) method used to extract a summary variable for each cluster (see details)}

\item{corr.method}{(character) which correlation coefficient to use: \code{c('pearson', 'spearman')}}

\item{corr.use}{(character) how to deal with missing values: \code{c('complete.obs', 'pairwise.complete.obs')}}

\item{clus.method}{(character) how to calculate linkages: \code{c('complete', 'single', 'centroid')}}
}
\value{
a list containing the following elements:
\itemize{
\item{\strong{nvar} - number of variables in \code{x}}
\item{\strong{nclust} - number of created clusters}
\item{\strong{vif} - VIF values for each original variable in \code{x}}
\item{\strong{clusters} - a list for each cluster containing the number of variables in the cluster, 
the names of the variables in the cluster, all pairwise correlations, and predictor-response correlations)}
\item{\strong{summaries} - a data frame containing the summary variables (one for each cluster)}
\item{\strong{corrplot} - a ggplot2 object visualizing the correlations amongst all predictor variables}
\item{\strong{params} - a list of all parameter values passed in the original call}
}
}
\description{
Perform agglomerative (hierarchical) clustering on a set of numeric predictor variables on the basis of 
absolute correlation distance. Create a summary variable for each cluster to use for dimension reduction.
}
\details{
Variables will be agglomerated into clusters on the basis of absolute correlation distance (1 - abs(cor)) 
using a hierarchical clustering algorithm (hclust). The resulting dendogram will be cut based on the value
of \code{corr.min} supplied by the user, such that all clusters contain variables with linkage correlations
of at least \code{corr.min}. A single summary variable will be created for each cluster based on the value
of \code{clus.summary} supplied by the user, with behavior detailed below. For all variables not linked into
a cluster, the original variable is returned as the "summary" measure unchanged.
\itemize{
\item{\strong{max.pw} - original variable with highest average pairwise correlation amongst cluster variables}
\item{\strong{max.y} - original variable with highest correlation with the response (\code{y})}
\item{\strong{avg.x} - average of standardized variables within the cluster for each observation}
\item{\strong{pc.1} - first principal component of cluster variables}
}
}
\examples{
library(caret)
data("ChemicalManufacturingProcess")
x <- ChemicalManufacturingProcess[,-1]
y <- ChemicalManufacturingProcess[, 1]
res <- varCluster(x, y, corr.min = 0.75, clus.summary = 'pc.1', corr.method = 'spearman', corr.use = 'complete.obs')
res$nvar
res$nclust
res$clusters
str(res$summaries)
res$corrrplot
res$params
}

